Authorization with policies

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -8,6 +8,7 @@ use App\Models\Attendee;
 use App\Models\Event;
 use Illuminate\Http\Request;
 use App\Http\Traits\CanLoadRelationships;
+use Illuminate\Support\Facades\Gate;
 
 class AttendeeController extends Controller
 {
@@ -66,6 +67,8 @@ class AttendeeController extends Controller
      */
     public function destroy(string $event, Attendee $attendee)
     {
+
+      Gate::authorize('delete-attendee',[$event,$attendee]);
       $attendee->delete();
